+++ app/main/db.py
@@ -53,6 +53,10 @@ class Entity():
         if not parent_entity_id:
             return entity_id
 
+        # Propagate sharing
+        parent = self.db.get('SELECT sharing FROM entity WHERE id = %s LIMIT 1;', parent_entity_id)
+        self.db.execute('UPDATE entity SET sharing = %s WHERE id = %s LIMIT 1;', parent.sharing, entity_id)
+
         # Insert child relationship and/or default parent child relationship
         sql = """
             INSERT INTO relationship (
@@ -106,10 +110,13 @@ class Entity():
             AND       r.related_entity_id = %s
             AND       r.relationship_definition_keyname = 'child'
             AND       rr.is_deleted = 0
-            AND       rr.relationship_definition_keyname IN ('viewer', 'editor', 'owner' );
+            AND       rr.relationship_definition_keyname IN ('viewer', 'expander', 'editor', 'owner');
         """
         # logging.debug(sql)
         self.db.execute(sql, entity_id, self.__user_id, entity_id)
+
+        # set creator to owner
+        self.set_rights(entity_id=entity_id, related_entity_id=self.__user_id, right='owner')
 
         # Populate default values
         for default_value in self.db.query('SELECT keyname, defaultvalue FROM property_definition WHERE entity_definition_keyname = %s AND defaultvalue IS NOT null', entity_definition_keyname):
@@ -203,6 +210,9 @@ class Entity():
             formula = Formula(self.db, user_locale=self.get_user_locale(), created_by=self.__user_id, entity_id=entity_id, property_id=new_property_id, formula=value)
             value = ''.join(formula.evaluate())
 
+        if definition.datatype != 'file':
+            value_string = value[:500]
+
         if definition.datatype in ['text', 'html']:
             field = 'value_text'
         elif definition.datatype == 'integer':
@@ -211,16 +221,19 @@ class Entity():
             field = 'value_decimal'
             value = value.replace(',', '.')
             value = re.sub(r'[^\.0-9:]', '', value)
+            value_string = '%s %s'[:500] % (value_string, value)
         elif definition.datatype == 'date':
             field = 'value_datetime'
         elif definition.datatype == 'datetime':
             field = 'value_datetime'
         elif definition.datatype == 'reference':
             field = 'value_reference'
+            value_string = self.__get_properties(value)[0]['displayname']
         elif definition.datatype == 'file':
             uploaded_file = value
             value = self.db.execute_lastrowid('INSERT INTO file SET filename = %s, filesize = %s, file = %s, created_by = %s, created = NOW();', uploaded_file['filename'], len(uploaded_file['body']), uploaded_file['body'], self.__user_id)
             field = 'value_file'
+            value_string = uploaded_file['filename'][:500]
         elif definition.datatype == 'boolean':
             field = 'value_boolean'
             value = 1 if value.lower() == 'true' else 0
@@ -229,10 +242,14 @@ class Entity():
         else:
             field = 'value_string'
             value = value[:500]
+            value_string = ''
 
         # logging.debug('UPDATE property SET %s = %s WHERE id = %s;' % (field, value, new_property_id) )
 
-        self.db.execute('UPDATE property SET %s = %%s WHERE id = %%s;' % field, value, new_property_id )
+        if value_string:
+            self.db.execute('UPDATE property SET %s = %%s, value_string = %%s WHERE id = %%s;' % field, value, value_string, new_property_id )
+        else:
+            self.db.execute('UPDATE property SET %s = %%s WHERE id = %%s;' % field, value, new_property_id )
 
         if definition.formula == 1:
             formula.save_property(new_property_id=new_property_id, old_property_id=old_property_id)
@@ -376,7 +393,6 @@ class Entity():
         # logging.debug(sql)
 
         property_id = self.db.execute_lastrowid(sql)
-        # logging.warning(str(property_id))
         return self.db.get('SELECT value_string FROM property WHERE id = %s', property_id).value_string
 
     def set_relations(self, entity_id, related_entity_id, relationship_definition_keyname, delete=False, update=False):
@@ -400,7 +416,7 @@ class Entity():
                     if delete == True:
                         sql = """
                             UPDATE relationship SET
-                                deleted_by = '%s',
+                                deleted_by = %s,
                                 deleted = NOW(),
                                 is_deleted = 1
                             WHERE relationship_definition_keyname = '%s'
@@ -414,7 +430,7 @@ class Entity():
                             UPDATE relationship SET
                                 deleted_by = NULL,
                                 deleted = NULL,
-                                changed_by = '%s',
+                                changed_by = %s,
                                 changed = NOW()
                             WHERE relationship_definition_keyname = '%s'
                             AND entity_id = %s
@@ -428,7 +444,7 @@ class Entity():
                                     relationship_definition_keyname = '%s',
                                     entity_id = %s,
                                     related_entity_id = %s,
-                                    created_by = '%s',
+                                    created_by = %s,
                                     created = NOW();
                             """ % (t, e, r, self.__user_id)
                             # logging.debug(sql)
@@ -439,11 +455,81 @@ class Entity():
                                 relationship_definition_keyname = '%s',
                                 entity_id = %s,
                                 related_entity_id = %s,
-                                created_by = '%s',
+                                created_by = %s,
                                 created = NOW();
                         """ % (t, e, r, self.__user_id)
                         # logging.debug(sql)
                         self.db.execute(sql)
+
+    def get_rights(self, entity_id):
+        if not entity_id:
+            return
+
+        rights = {}
+        for right in ['viewer', 'expander', 'editor', 'owner']:
+            sql = """
+                SELECT related_entity_id
+                FROM relationship
+                WHERE is_deleted = 0
+                AND relationship_definition_keyname = %s
+                AND entity_id = %s
+            """
+
+            relationship = self.db.query(sql, right, entity_id)
+            if not relationship:
+                continue
+
+            ids = [x.related_entity_id for x in relationship if x.related_entity_id]
+            if ids:
+                entities = self.__get_properties(entity_id=ids, full_definition=False, only_public=False)
+                if entities:
+                    rights[right] = entities
+
+        return rights
+
+    def set_rights(self, entity_id, related_entity_id, right=None):
+        if not entity_id or not related_entity_id:
+            return
+
+        if type(entity_id) is not list:
+            entity_id = [entity_id]
+
+        if type(related_entity_id) is not list:
+            related_entity_id = [related_entity_id]
+
+        sql = """
+            UPDATE relationship SET
+                deleted = NOW(),
+                is_deleted = 1,
+                deleted_by = %s
+            WHERE is_deleted = 0
+            AND relationship_definition_keyname IN ('viewer', 'expander', 'editor', 'owner')
+            AND entity_id IN (%s)
+            AND related_entity_id IN (%s);
+        """ % (self.__user_id, ','.join(map(str, entity_id)), ','.join(map(str, related_entity_id)))
+        self.db.execute(sql)
+
+        if right in ['viewer', 'expander', 'editor', 'owner']:
+            for e in entity_id:
+                for re in related_entity_id:
+                    self.db.execute('INSERT INTO relationship SET relationship_definition_keyname = %s, entity_id = %s, related_entity_id = %s, created = NOW(), created_by = %s;', right, int(e), int(re), self.__user_id)
+
+    def set_sharing(self, entity_id, sharing):
+        if not entity_id or not sharing:
+            return
+
+        if type(entity_id) is not list:
+            entity_id = [entity_id]
+
+        sql = """
+            UPDATE entity SET
+                sharing = %%s,
+                changed = NOW(),
+                changed_by = %%s
+            WHERE id IN (%s);
+        """ % ','.join(map(str, entity_id))
+
+        self.db.execute(sql, sharing, self.__user_id)
 
     def get_entities(self, ids_only=False, entity_id=None, search=None, entity_definition_keyname=None, dataproperty=None, limit=None, full_definition=False, only_public=False):
         """
@@ -500,9 +586,9 @@ class Entity():
         if self.__user_id and only_public == False:
             where_parts.append('r.is_deleted = 0')
             join_parts.append('RIGHT JOIN relationship AS r  ON r.entity_id  = e.id')
-            where_parts.append('r.related_entity_id IN (%s) AND r.relationship_definition_keyname IN (\'viewer\', \'editor\', \'owner\')' % self.__user_id)
+            where_parts.append('(r.related_entity_id = %s AND r.relationship_definition_keyname IN (\'viewer\', \'expander\', \'editor\', \'owner\') OR e.sharing = \'domain\')' % self.__user_id)
         else:
-            where_parts.append('e.public = 1')
+            where_parts.append('e.sharing = \'public\'')
             i = 0
             if search != None:
                 for s in search.split(' '):
@@ -558,9 +644,13 @@ class Entity():
                 entity_id = [entity_id]
 
             if self.__user_id and only_public == False:
+                rights_select = 'relationship.relationship_definition_keyname'
+                rights_join = 'LEFT JOIN relationship ON relationship.entity_id = entity.id AND relationship.is_deleted = 0 AND relationship.related_entity_id = %s AND relationship.relationship_definition_keyname IN (\'viewer\', \'expander\', \'editor\', \'owner\')' % self.__user_id
                 public = ''
             else:
-                public = 'AND entity.public = 1 AND property_definition.public = 1'
+                rights_select = 'FALSE'
+                rights_join = ''
+                public = 'AND entity.sharing = \'public\' AND property_definition.public = 1'
 
             datapropertysql = ''
             if dataproperty:
@@ -577,7 +667,8 @@ class Entity():
                     entity_definition.%(language)s_description      AS entity_description,
                     entity.created                                  AS entity_created,
                     entity.changed                                  AS entity_changed,
-                    entity.public                                   AS entity_public,
+                    entity.sharing                                  AS entity_sharing,
+                    %(rights_select)s                               AS entity_right,
                     entity_definition.%(language)s_displayname      AS entity_displayname,
                     entity_definition.%(language)s_displayinfo      AS entity_displayinfo,
                     entity_definition.%(language)s_displaytable     AS entity_displaytable,
@@ -611,7 +702,7 @@ class Entity():
                     property.value_reference                        AS value_reference,
                     property.value_file                             AS value_file
                 FROM
-                    entity,
+                    entity %(rights_join)s,
                     entity_definition,
                     property,
                     property_definition
@@ -628,11 +719,16 @@ class Entity():
                 ORDER BY
                     entity_definition.keyname,
                     entity.created DESC
-            """ % {'language': self.get_user_locale().code, 'public': public, 'idlist': ','.join(map(str, entity_id)), 'datapropertysql': datapropertysql}
+            """ % {'language': self.get_user_locale().code, 'public': public, 'idlist': ','.join(map(str, entity_id)), 'datapropertysql': datapropertysql, 'rights_select': rights_select, 'rights_join': rights_join}
             # logging.debug(sql)
 
             items = {}
             for row in self.db.query(sql):
+                if row.entity_sharing == 'private' and not row.entity_right:
+                    continue
+                if row.entity_sharing == 'domain' and row.entity_right not in ['viewer', 'expander', 'editor', 'owner'] and row.property_public != 1:
+                    continue
+
                 #Entity
                 items.setdefault('item_%s' % row.entity_id, {})['definition_keyname'] = row.entity_definition_keyname
                 items.setdefault('item_%s' % row.entity_id, {})['id'] = row.entity_id
@@ -647,7 +743,8 @@ class Entity():
                 items.setdefault('item_%s' % row.entity_id, {})['displayinfo'] = row.entity_displayinfo
                 items.setdefault('item_%s' % row.entity_id, {})['displaytable'] = row.entity_displaytable
                 items.setdefault('item_%s' % row.entity_id, {})['file_count'] = 0
-                items.setdefault('item_%s' % row.entity_id, {})['is_public'] = True if row.entity_public == 1 else False
+                items.setdefault('item_%s' % row.entity_id, {})['sharing'] = row.entity_sharing
+                items.setdefault('item_%s' % row.entity_id, {})['right'] = row.entity_right
                 items.setdefault('item_%s' % row.entity_id, {})['ordinal'] = row.entity_created if row.entity_created else datetime.datetime.now()
 
                 #Property
@@ -724,16 +821,16 @@ class Entity():
                     value = row.value_text if row.value_text else ''
                 elif row.property_datatype == 'integer':
                     db_value = row.value_integer
-                    value = row.value_integer
+                    value = row.value_integer if row.value_integer else ''
                 elif row.property_datatype == 'decimal':
                     db_value = row.value_decimal
-                    value = '%.2f' % row.value_decimal
+                    value = '%.2f' % row.value_decimal if row.value_decimal else ''
                 elif row.property_datatype == 'date':
                     db_value = row.value_datetime
                     value = formatDatetime(row.value_datetime, '%(day)02d.%(month)02d.%(year)d')
                 elif row.property_datatype == 'datetime':
                     db_value = row.value_datetime
-                    value = formatDatetime(row.value_datetime)
+                    value = formatDatetime(row.value_datetime) if row.value_datetime else ''
                 elif row.property_datatype == 'reference':
                     value = ''
                     if row.value_reference:
@@ -811,7 +908,7 @@ class Entity():
                     else:
                         items[key].setdefault('properties', {}).setdefault('%s' % d.property_dataproperty, {})['can_add_new'] = False
 
-                    if d.property_classifier_id:
+                    if d.property_classifier_id and d.property_datatype != 'reference':
                         for c in self.get_entities(entity_definition_keyname=d.property_classifier_id, only_public=True):
                             if c.get('id', None):
                                 items[key].setdefault('properties', {}).setdefault('%s' % d.property_dataproperty, {}).setdefault('select', []).append({'id': c.get('id', ''), 'label': c.get('displayname', '')})
@@ -1073,15 +1170,15 @@ ORDER BY child.sort, child.created DESC;"""
             sql += ' AND r.related_entity_id IN (%s)' % ','.join(map(str, related_entity_id))
 
         if self.__user_id and only_public == False:
-            sql += ' AND rights.related_entity_id IN (%s) AND rights.relationship_definition_keyname IN (\'viewer\', \'editor\', \'owner\')' % self.__user_id
+            sql += ' AND (rights.related_entity_id = %s AND rights.relationship_definition_keyname IN (\'viewer\', \'expander\', \'editor\', \'owner\') OR e.sharing = \'domain\')' % self.__user_id
         else:
-            sql += ' AND e.public = 1'
+            sql += ' AND e.sharing = \'public\''
 
         if relationship_definition_keyname:
             sql += ' AND r.relationship_definition_keyname IN (%s)' % ','.join(['\'%s\'' % x for x in relationship_definition_keyname])
 
         if entity_definition_keyname:
-            sql += ' AND e.entity_definition_keyname IN (\'%s\')' % ','.join(map(str, entity_definition_keyname))
+            sql += ' AND e.entity_definition_keyname IN (%s)' % ','.join(['\'%s\'' % x for x in entity_definition_keyname])
 
         if unionsql:
             sql += unionsql
@@ -1090,12 +1187,12 @@ ORDER BY child.sort, child.created DESC;"""
                 sql += ' AND up.value_reference IN (%s)' % ','.join(map(str, entity_id))
 
             if self.__user_id and only_public == False:
-                sql += ' AND urights.related_entity_id IN (%s) AND urights.relationship_definition_keyname IN (\'viewer\', \'editor\', \'owner\')' % self.__user_id
+                sql += ' AND (urights.related_entity_id = %s AND urights.relationship_definition_keyname IN (\'viewer\', \'expander\', \'editor\', \'owner\') OR ue.sharing = \'domain\')' % self.__user_id
             else:
                 sql += ' AND ue.public = 1'
 
             if entity_definition_keyname:
-                sql += ' AND ue.entity_definition_keyname IN (\'%s\')' % ','.join(map(str, entity_definition_keyname))
+                sql += ' AND ue.entity_definition_keyname IN (%s)' % ','.join(['\'%s\'' % x for x in entity_definition_keyname])
 
         sql += ' ORDER BY sort, created DESC'
 
@@ -1189,6 +1286,10 @@ ORDER BY child.sort, child.created DESC;"""
         Returns allowed child definitions.
 
         """
+
+        if not self.db.get('SELECT id FROM relationship WHERE relationship_definition_keyname iN (\'expander\', \'owner\') AND entity_id = %s AND related_entity_id = %s LIMIT 1;', entity_id, self.__user_id):
+            return []
+
         sql = """
             SELECT DISTINCT
                 ed.keyname,
@@ -1202,7 +1303,8 @@ ORDER BY child.sort, child.created DESC;"""
             WHERE r.relationship_definition_keyname = 'allowed-child'
             AND r.entity_id = %(id)s
             AND r.is_deleted = 0
-            ORDER BY ed.keyname        """  % {'language': self.get_user_locale().code, 'id': entity_id}
+            ORDER BY ed.keyname
+        """  % {'language': self.get_user_locale().code, 'id': entity_id}
         # logging.debug(sql)
 
         result = self.db.query(sql)
@@ -1277,9 +1379,9 @@ ORDER BY child.sort, child.created DESC;"""
                 relationship
             WHERE entity.entity_definition_keyname = entity_definition.keyname
             AND relationship.entity_id = entity.id
-            AND relationship.relationship_definition_keyname IN ('viewer', 'editor', 'owner')
+            AND relationship.relationship_definition_keyname IN ('viewer', 'expander', 'editor', 'owner')
             AND entity_definition.estonian_menu IS NOT NULL
-            AND relationship.related_entity_id IN (%(user_id)s)
+            AND relationship.related_entity_id = %(user_id)s
             AND entity.is_deleted = 0
             AND relationship.is_deleted = 0
             ORDER BY
@@ -1733,8 +1835,8 @@ def mdbg(matchobj):
     # mdbg() is for regex match object debugging.
     #   i.e: re.sub(r"([^{]*){([^{}]*)}|(.*?)$", mdbg, self.formula)(ha()a)
     for m in matchobj.groups():
-        None
-        logging.debug(m)
+        pass
+        # logging.debug(m)
 
 
 def sortableDateTime(s_date):