+++ HISTORY.md
@@ -1,3 +1,13 @@
+## 2013-05-01
+* Show and allow edit Entities based on sharing and rights settings
+* Additional right (to viewer, editor, owner) is right to add child Entities
+* Bugfixes
+
+## 2013-04-29
+* View and set Entity sharing and rights
+* Save all request info to app_requests table
+* New 404 page
+
 ## 2013-03-22
 * Host based database selection - one process can serve multiple hosts/databases
 * Reference propery values are now links to referred entity